runpod-workers/ddcolor/requirements.txt
runpod>=1.6.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
modelscope[cv]>=1.10.0
torch>=2.0.0
torchvision>=0.15.0
